move defaults to where the data models are to have them somewhere central
